requirements.txt
discord.py==2.3.2
yt-dlp==2023.10.22
asyncio==3.4.3
ffmpeg-python==0.2.0
PyNaCl==1.5.0
